Update scodec from version 1.0.4 to 1.0.6. This also changes the group id. Specs2 version 2.4.9 depended on scodec 1.0.2, so it was upgraded to Specs2 3.5.

+++ dynamodb/src/test/scala/io/atlassian/aws/dynamodb/DecodeSpec.scala
@@ -43,7 +43,7 @@ class DecodeSpec extends ScalaCheckSpec {
     (Encoder[Int].encode(100) |> Decoder[NonEmptyBytes].decode).toOr.toEither must beLeft
 
   def mapAttemptPropagatesExceptions =
-    (Encoder[NonEmptyBytes].encode(ByteVector.fromByte(1) match {
+    (Encoder[NonEmptyBytes].encode(ByteVector.fromByte(1.toByte) match {
       case NonEmptyBytes(b) => b
     }) |> Decoder[TwoLongs].decode).toOr.toEither must beLeft
 }
